requirements.txt
# Core PyTorch - install first
torch>=2.0.0
torchvision
torchaudio

# Neo4j and Graph Database
neo4j>=5.0.0
py2neo>=2021.2.0

# Data Processing and Analysis - install before PyTorch Geometric
pandas>=1.5.0
numpy>=1.20.0
networkx>=2.8.0
scikit-learn>=1.2.0

# PyTorch Geometric ecosystem - install after torch
torch-geometric>=2.3.0
# Note: torch-scatter and torch-sparse will be installed automatically with torch-geometric

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.64.0
python-dotenv>=1.0.0
requests>=2.28.0

# Development
jupyter>=1.0.0
ipykernel>=6.20.0
